requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.2,<3.0
python-telegram-bot>=21.0,<22.0
redis>=5.0,<6.0
celery[redis]>=5.4,<6.0
flower>=2.0,<3.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0

